+++ crates/testing-sdk/tests/sdk/plain_rolling.rs
@@ -172,7 +172,7 @@ pub(crate) fn input_row(rn: u64, group: &str, ts: u64, value: f64) -> CoreRow {
 	TestOperatorRowBuilder::new(rn)
 		.with_values(vec![Value::Utf8(group.into()), Value::Uint8(ts), Value::float8(value)])
 		.with_fields(input_fields())
-		.with_time(DateTime::from_millis(ts))
+		.with_time(DateTime::from_millis(ts as i64))
 		.build()
 }
