+++ crates/testing-sdk/tests/sdk/plain_sliding.rs
@@ -29,7 +29,7 @@ fn sliding(size: u64, slide: u64, lateness: u64) -> ApplyWith {
 }
 
 fn at(ms: u64) -> u64 {
-	DateTime::from_millis(ms).to_order()
+	DateTime::from_millis(ms as i64).to_order()
 }
 
 #[test]
